requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
psycopg2-binary==2.9.6
boto3==1.28.24

